Extract evaluate_planet and helper functions to the separate files
