+++ default/python/AI/InvasionAI.py
@@ -12,6 +12,9 @@ import MilitaryAI
 import PlanetUtilsAI
 from AIDependencies import INVALID_ID, Tags
 from aistate_interface import get_aistate
+from colonization.calculate_planet_colonization_rating import (
+    calculate_planet_colonization_rating,
+)
 from common.print_utils import Number, Table, Text
 from EnumsAI import MissionType, PriorityType
 from freeorion_tools import (
@@ -353,11 +356,21 @@ def evaluate_invasion_planet(planet_id, secure_fleet_missions, verbose=True):
         # this call iterates over this Empire's available species with which it could colonize after an invasion
         planet_eval = ColonisationAI.assign_colonisation_values([planet_id], MissionType.INVASION, None, detail)
         colony_base_value = max(0.75 * planet_eval.get(planet_id, [0])[0],
-                                ColonisationAI.evaluate_planet(
-                                    planet_id, MissionType.OUTPOST, None, detail, empire_research_list))
+                                calculate_planet_colonization_rating(
+                                    planet_id=planet_id,
+                                    mission_type=MissionType.OUTPOST,
+                                    spec_name=None,
+                                    detail=detail,
+                                    empire_research_list=empire_research_list
+                                ))
     else:
-        colony_base_value = ColonisationAI.evaluate_planet(
-            planet_id, MissionType.INVASION, species_name, detail, empire_research_list)
+        colony_base_value = calculate_planet_colonization_rating(
+            planet_id=planet_id,
+            mission_type=MissionType.INVASION,
+            spec_name=species_name,
+            detail=detail,
+            empire_research_list=empire_research_list
+        )
 
     # Add extra score for all buildings on the planet
     building_values = {"BLD_IMPERIAL_PALACE": 1000,
