+++ default/python/AI/colonization/colony_score.py
@@ -1,0 +1,2 @@
+# minimum evaluation score that a planet must reach so it is considered for outposting or colonizing
+MINIMUM_COLONY_SCORE = 60
